Update to latest leap exercise

+++ rust/leap/src/lib.rs
@@ -1,4 +1,4 @@
-pub fn is_leap_year(year: i32) -> bool {
+pub fn is_leap_year(year: u64) -> bool {
     if year % 400 == 0 {
         true
     } else if year % 100 == 0 {
